revamp stages 3-4: dialect API and the instrument package
dialect.py is the public extension surface: register_ufunc,
register_function, register_opaque, register_neutral,
register_contract, each documented with the house rule (names
propose, runs dispose -- ship a differential test with every row).

instrument.py becomes a package of named subsystems: registries
(the declared roles, one line per fact), runtime (the traced
twins), triage (the runtime dispatch policy as one ordered rule
list -- the artist formerly known as doorman), rewriter (the AST
visitor), twins (building instrumented functions, classes and
methods, caches on the session). Mutually recursive seams
(triage <-> twins, runtime -> triage) resolve at call time and say
so. The package __init__ re-exports every historical name; no
caller or test changed.
